pyproject.toml
[project]
name = "intake-ai-cli"
version = "0.2.0"
description = "From requirements in any format to verified implementation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [{name = "intake contributors"}]
keywords = ["spec-driven", "requirements", "ai", "cli", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    # CLI
    "click>=8.1",
    "rich>=13.0",

    # Configuration and validation
    "pydantic>=2.0",
    "pyyaml>=6.0",

    # LLM
    "litellm>=1.40",

    # Parsers
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "beautifulsoup4>=4.12",
    "markdownify>=0.13",

    # HTTP (for future URL/API integrations)
    "httpx>=0.27",

    # Templates
    "jinja2>=3.1",

    # Logging
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "types-beautifulsoup4>=4.12",
]
connectors = [
    "atlassian-python-api>=3.40",
    "PyGithub>=2.0",
]
watch = [
    "watchfiles>=0.21",
]
mcp = [
    "mcp>=1.0",
]

[project.scripts]
intake = "intake.cli:main"

[project.entry-points."intake.parsers"]
markdown = "intake.ingest.markdown:MarkdownParser"
plaintext = "intake.ingest.plaintext:PlaintextParser"
yaml = "intake.ingest.yaml_input:YamlInputParser"
pdf = "intake.ingest.pdf:PdfParser"
docx = "intake.ingest.docx:DocxParser"
jira = "intake.ingest.jira:JiraParser"
confluence = "intake.ingest.confluence:ConfluenceParser"
image = "intake.ingest.image:ImageParser"
url = "intake.ingest.url:UrlParser"
slack = "intake.ingest.slack:SlackParser"
github_issues = "intake.ingest.github_issues:GithubIssuesParser"

[project.entry-points."intake.exporters"]
architect = "intake.export.architect:ArchitectExporter"
generic = "intake.export.generic:GenericExporter"

[project.entry-points."intake.connectors"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/intake"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"
asyncio_mode = "auto"

[project.urls]
Homepage = "https://diego303.github.io/intake-docs/"
Repository = "https://github.com/Diego303/intake-cli"